+++ src/www/definitions.ts
@@ -50,6 +50,8 @@ export interface WebViewOptions {
   showNavigationButtons: boolean;
   leftToRight: boolean;
   
+  customWebViewUserAgent: String | null;
+
   android: AndroidWebViewOptions,
   iOS: iOSWebViewOptions
 }